Fixes encoding of schema that are just strings
In encoding arrays for instance, the "items" (defining a type)
can just be a name (primitive), and we shouldn't try to
call type() on this string to get the actual type.

Also implements all examples in the spec. Now I need to shuffle
around to do schema type resolution on input data (kind of yucky)